#7 #10 selenium drivers - don't fail uninstall if exe does not exist - selenium drivers - stop use deprecated API (Get-BinRoot)

+++ manual/selenium-opera-driver/x86_64/tools/chocolateyUninstall.ps1
@@ -1,23 +1,29 @@
-$binRoot = Get-BinRoot
-$seleniumDir = "$binRoot\selenium"
-$driverPath = "$seleniumDir\operadriver.exe"
+$toolsLocation = Get-ToolsLocation
+$seleniumDir = "$toolsLocation\selenium"
 
-Remove-Item $driverPath -Force
+if (Test-Path $seleniumDir) {
+  $driverPath = "$seleniumDir\operadriver.exe"
+  if (Test-Path $driverPath) {
+    Remove-Item $driverPath -Force
+  }
 
-$directoryInfo = Get-ChildItem $seleniumDir | Measure-Object
-If ($directoryInfo.count -eq 0) {
-  Remove-Item $seleniumDir -Force
+  $directoryInfo = Get-ChildItem $seleniumDir | Measure-Object
+  If ($directoryInfo.count -eq 0) {
+    Remove-Item $seleniumDir -Force
+  }
 }
 
 $menuPrograms = [environment]::GetFolderPath([environment+specialfolder]::Programs)
 $shortcutDir = "$menuPrograms\Selenium"
-$shortcutFile = "$shortcutDir\Selenium Opera Driver.lnk"
 
-If (Test-Path $shortcutFile) {
-  Remove-Item $shortcutFile -Force
-}
+if (Test-Path $shortcutDir) {
+  $shortcutFile = "$shortcutDir\Selenium Opera Driver.lnk"
+  If (Test-Path $shortcutFile) {
+    Remove-Item $shortcutFile -Force
+  }
 
-$directoryInfo = Get-ChildItem $shortcutDir | Measure-Object
-If ($directoryInfo.count -eq 0) {
-  Remove-Item $shortcutDir -Force
+  $directoryInfo = Get-ChildItem $shortcutDir | Measure-Object
+  If ($directoryInfo.count -eq 0) {
+    Remove-Item $shortcutDir -Force
+  }
 }
